# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Watch Clouds - Movie & Web Series Streaming Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tmdb/search:
    get:
      operationId: tmdbSearch
      tags: [tmdb]
      summary: Search TMDB for movies/series
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [movie, tv]
            default: movie
      responses:
        "200":
          description: TMDB search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TmdbResult"
  /tmdb/import:
    post:
      operationId: tmdbImport
      tags: [tmdb]
      summary: Import movie/series from TMDB to database (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TmdbImportInput"
      responses:
        "201":
          description: Imported from TMDB
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movie"

  # Watchlist
...
  /watchlist/{movieId}:
    post:
      operationId: addToWatchlist
      tags: [watchlist]
      summary: Add movie to watchlist
      parameters:
        - name: movieId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Added to watchlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

    delete:
      operationId: removeFromWatchlist
      tags: [watchlist]
      summary: Remove from watchlist
      parameters:
        - name: movieId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Removed from watchlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Reviews
  /movies/{id}/reviews:
    get:
      operationId: getMovieReviews
      tags: [reviews]
      summary: Get reviews for a movie
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reviews list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"

    post:
      operationId: createReview
      tags: [reviews]
      summary: Add a review for a movie
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # Admin dashboard stats
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Movie:
      type: object
      required: [id, title, type, posterUrl, year, rating, genres]
      properties:
        id:
          type: string
        title:
          type: string
        type:
          type: string
          enum: [movie, webseries, anime, documentary]
        posterUrl:
          type: string
        backdropUrl:
          type: ["string", "null"]
        year:
          type: integer
        rating:
          type: number
        genres:
          type: array
          items:
            type: string
        overview:
          type: ["string", "null"]
        tmdbId:
          type: ["integer", "null"]
        isFeatured:
          type: boolean
        streamUrl:
          type: ["string", "null"]
        trailerUrl:
          type: ["string", "null"]
        runtime:
          type: ["integer", "null"]
        language:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    TmdbImportInput:
      type: object
      required: [tmdbId, mediaType]
      properties:
        tmdbId:
          type: integer
        mediaType:
          type: string
          enum: [movie, tv]
    TmdbResult:
      type: object
      required: [tmdbId, title, type, posterUrl, year, rating]
      properties:
        tmdbId:
          type: integer
        title:
          type: string
        type:
          type: string
        posterUrl:
          type: string
        backdropUrl:
          type: ["string", "null"]
        year:
          type: integer
        rating:
          type: number
        overview:
          type: ["string", "null"]
    Review:
      type: object
      required: [id, userId, movieId, rating, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        userName:
          type: string
        movieId:
          type: string
        rating:
          type: number
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ReviewInput:
      type: object
      required: [rating]
      properties:
        rating:
          type: number
          minimum: 1
          maximum: 10
        comment:
          type: string
...